fix(eventServiceV2): migrated createEventV2 to child class
