+++ apps/backend/src/Events/dto/EventDto.dto.ts
@@ -59,6 +59,24 @@ export class EventCriteriaDto {
   @ApiPropertyOptional() @IsOptional() @IsUUID() moduleId?: string;
 }
 
+export class EventCriteriaDtoV2 extends PartialType(
+  PickType(EventCriteriaDto, [
+    'date',
+    'dayOfWeek',
+    'startTime',
+    'endTime',
+    'moduleId',
+  ]),
+) {
+  @ApiProperty({
+    default: false,
+    description: `moduleId required in V2 - whomp whomp`,
+    nullable: false,
+  })
+  @IsUUID()
+  moduleId!: string;
+}
+
 export class UpdateEventCriteriaDto extends PartialType(EventCriteriaDto) {}
 
 export class VenueDto {
@@ -139,6 +157,21 @@ export class CreateEventDto extends PickType(EventDto, [
   @IsEnum(ActivityTypeSchema.options)
   activityType?: (typeof ActivityTypeSchema.options)[number];
 }
+
+export class CreateEventDtoV2 extends PickType(EventDto, [
+  'eventName',
+  'activityCode',
+  'activityType',
+  'venues',
+  'isRecurring',
+  'validated',
+]) {
+  @ApiProperty({ type: EventCriteriaDtoV2 })
+  @ValidateNested()
+  @Type(() => EventCriteriaDtoV2)
+  eventCriteria!: EventCriteriaDtoV2;
+}
+
 export class UpdateEventDto extends PartialType(
   OmitType(EventDto, ['eventId', 'eventCriteria', 'venues'] as const),
 ) {
