+++ apps/backend/src/Events/event.controller.ts
@@ -18,26 +18,44 @@ import {
   UpdateEventDto,
   DeleteResponseDto,
   EventFiltersDto,
+  CreateEventDtoV2,
 } from './dto/EventDto.dto';
 
 import { EventService } from './event.service';
 import type { SessionData } from '../auth/session.decorator';
 import { CurrentSession } from '../auth/session.decorator';
 import { Roles } from '../auth/roles.guard';
+import { EventServiceV2 } from './eventV2.service';
 
 @ApiTags('Events')
 @Controller('events')
 export class EventController {
-  constructor(private readonly service: EventService) {}
+  constructor(
+    private readonly service: EventService,
+    private readonly service2: EventServiceV2,
+  ) {}
 
-  @Post('/v2')
+  @Post('v2')
   @Roles('lecturer', 'uni_admin')
-  @ApiBody({ type: CreateEventDto })
+  @ApiOperation({
+    summary: 'Create an event - V2',
+    operationId: 'createEventV2',
+  })
+  @ApiBody({ type: CreateEventDtoV2 })
+  @ApiResponse({
+    status: 201,
+    description: 'Event created successfully',
+    type: EventSingleResponseDto,
+  })
   createEvent2(
     @CurrentSession() session: SessionData,
-    @Body() dto: CreateEventDto,
+    @Body() dto: CreateEventDtoV2,
   ) {
-    return this.service.createV2(dto, session.user.id, session.uniId);
+    //make activityType and activityCode very optional
+    //isrecurring default to false - just need date
+    //validated - dfault false - need a validate event endpoint!!!!!!!!!
+    //
+    return this.service2.createV2(dto, session.user.id, session.uniId);
   }
 
   @Post()
