+++ apps/backend/src/Events/event.module.ts
@@ -6,11 +6,22 @@ import { EventImportFingerprintService } from './event-import-fingerprint.servic
 import { ModuleModule } from 'src/Module/module.module';
 import { CourseModule } from 'src/Course/course.module';
 import { UniversityService } from 'src/University/university.service';
+import { EventServiceV2 } from './eventV2.service';
 
 @Module({
   imports: [forwardRef(() => ModuleModule), CourseModule],
   controllers: [EventController],
-  providers: [EventService, EventImportFingerprintService, UniversityService],
-  exports: [EventService, EventImportFingerprintService, UniversityService],
+  providers: [
+    EventService,
+    EventServiceV2,
+    EventImportFingerprintService,
+    UniversityService,
+  ],
+  exports: [
+    EventService,
+    EventServiceV2,
+    EventImportFingerprintService,
+    UniversityService,
+  ],
 })
 export class EventModule {}
