+++ apps/backend/src/Testing/Factories/event.factory.ts
@@ -9,6 +9,8 @@ import {
 import { EventCriteria, EventSource } from '../../Events/dto/event.types';
 import {
   CreateEventDto,
+  CreateEventDtoV2,
+  EventCriteriaDtoV2,
   EventDto,
   EventSingleResponseDto,
 } from '../../Events/dto/EventDto.dto';
@@ -35,6 +37,21 @@ export function createEventCriteria(
 
   return { ...base, ...overrides };
 } //END_createEventCriteria
+
+export function createEventCriteriaDtoV2(
+  overrides: Partial<EventCriteriaDtoV2> = {},
+): EventCriteriaDtoV2 {
+  const base: EventCriteriaDtoV2 = {
+    date: '2026-01-12',
+    startTime: '08:30',
+    endTime: '10:20',
+    moduleId: randomUUID(),
+
+    ...overrides,
+  };
+
+  return { ...base, ...overrides };
+} //END_createEventCriteriaV2
 
 type EventEntity = typeof Event.$inferSelect;
 export function createEvent(
@@ -109,6 +126,17 @@ export function createCreateEventDto(event: EventEntity): CreateEventDto {
     validated: event.validated,
   };
 } ///END_createCreateEventDto
+
+export function createCreateEventDtoV2(event: EventEntity): CreateEventDtoV2 {
+  return {
+    eventName: event.eventName,
+    activityCode: event.activityCode ?? undefined,
+    activityType: event.activityType as CreateEventDto['activityType'],
+    eventCriteria: event.eventCriteria as EventCriteriaDtoV2,
+    isRecurring: event.isRecurring,
+    validated: event.validated,
+  };
+} ///END_createCreateEventDtoV2
 
 type UniversityEvent = typeof UniversityEvent.$inferSelect;
 export function createUniversityEvent(
